refact: merge util.database to core.database
